Fixed unit tests and commented out functions that weren't implemented or couldn't be validated with unit test Will need to go back and validate if function needs to be used in future.

+++ src/main/java/org/boris/expr/ExprDate.java
@@ -16,7 +16,7 @@ import java.util.Date;
 public class ExprDate extends ExprDecimal
 {
     public ExprDate(Date value) {
-        super(ExprType.Date, Double.toString(value.getTime()));
+        super(ExprType.Date, Long.toString(value.getTime()));
         this.cancelEvalOnMissing = true;
     }
